More build system work, this should compile now
