+++ .travis.yml
@@ -14,14 +14,12 @@ env:
     INSTALL_DIR=$TRAVIS_BUILD_DIR/install
     VERILATOR_ROOT=$INSTALL_DIR
     PATH=$PATH:$VERILATOR_ROOT/bin:$TRAVIS_BUILD_DIR/utils/bin
-    CHISEL3=3.2-SNAPSHOT
-    CHISEL_TESTERS=1.3-SNAPSHOT
 
 install:
-  # Install dependencies
   - bash .install_verilator.sh
   - verilator --version
 
 script:
-  - set -o pipefail && sbt -Dchisel3Version=$CHISEL3 -Dchisel-iotestersVersion=$CHISEL_TESTERS test
+  - set -o pipefail && sbt awl/test
+  - set -o pipefail && sbt rocket-awl/test
