+++ build.sbt
@@ -32,11 +32,10 @@ def javacOptionsVersion(scalaVersion: String): Seq[String] = {
   }
 }
 
-// Provide a managed dependency on X if -DXVersion="" is supplied on the command line.
 val defaultVersions = Map(
   "chisel3" -> "3.2-SNAPSHOT",
   "chisel-iotesters" -> "1.3-SNAPSHOT",
-  "rocketchip" -> "1.2-SNAPSHOT",
+  "rocketchip" -> "1.2-031419-SNAPSHOT",
 )
 
 name := "awl"
@@ -90,17 +89,13 @@ val awlSettings = Seq(
 // when compiling tests under 2.11.12
 // An explicit dependency on junit seems to alleviate this.
   libraryDependencies ++= Seq(
-    "org.typelevel" %% "spire" % "0.14.1",
-    "org.scalanlp" %% "breeze" % "0.13.2",
-    "junit" % "junit" % "4.12" % "test",
-    "org.scalatest" %% "scalatest" % "3.0.5" % "test",
-    "org.scalacheck" %% "scalacheck" % "1.14.0" % "test"
+    "org.scalatest" %% "scalatest" % "3.0.5" % "test"
   ),
 )
 
 val rocketSettings = Seq(
     name := "rocket-awl",
-    libraryDependencies ++= Seq("chisel3", "chisel-iotesters", "rocketchip").map {
+    libraryDependencies ++= Seq("chisel-iotesters", "rocketchip").map {
       dep: String => "edu.berkeley.cs" %% dep % sys.props.getOrElse(dep + "Version", defaultVersions(dep))
     },
     Test / parallelExecution := false,
@@ -127,3 +122,4 @@ val `rocket-awl` = (project in file("rocket")).
   settings(commonSettings: _*).
   settings(rocketSettings: _*).
   dependsOn(awl)
+
