--- rocket/.install_verilator.sh
@@ -1,1 +1,0 @@
-../.install_verilator.sh
