+++ src/main/scala/serial/TestUtils.scala
@@ -1,4 +1,4 @@
-package awl.test.serial
+package awl.serial
 
 import chisel3._
 import chisel3.core.IntParam
