+++ src/test/scala/serial/Encoding8b10bTest.scala
@@ -1,4 +1,4 @@
-package awl.test.serial
+package awl.serial
 
 import chisel3._
 import chisel3.util._
